Cleanup `stats` code
- remove unused ignoreModules param
- add BuckImpl+Stats
